List the target names in the Goto Symbol list (ctrl/cmd+r)
This is quite hacky way to parse the values of the 'target_name'
key and assign them a scope that will be listed by the "Goto Symbol"
command in Sublime. There doesn't really seem to be a way to do it
cleanly from the language definition file. For example the problem
with current solution is that everything within
the 'target_name': 'foo' part is not assigned usual scopes
(punctuation, string).

But it's quite handy nevertheless so it might be worth to diverge.